+++ src/main/java/joshie/harvest/core/base/tile/TileFillable.java
@@ -2,7 +2,6 @@ package joshie.harvest.core.base.tile;
 
 import joshie.harvest.api.ticking.IDailyTickable;
 import joshie.harvest.core.helpers.MCServerHelper;
-import joshie.harvest.core.network.PacketHandler;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 
@@ -27,7 +26,7 @@ public abstract class TileFillable extends TileDaily implements IDailyTickable {
     public void saveAndRefresh() {
         MCServerHelper.markForUpdate(worldObj, getPos(), 3);
         if (!worldObj.isRemote) {
-            PacketHandler.sendRefreshPacket(this);
+            MCServerHelper.markTileForUpdate(this);
         }
 
         markDirty();