+++ src/main/java/joshie/harvest/core/network/PacketHandler.java
@@ -1,11 +1,7 @@
 package joshie.harvest.core.network;
 
-import joshie.harvest.core.base.tile.TileHarvest;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.EntityPlayerMP;
-import net.minecraft.network.Packet;
-import net.minecraft.server.management.PlayerChunkMapEntry;
-import net.minecraft.world.WorldServer;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
 import net.minecraftforge.fml.relauncher.Side;
 
@@ -42,23 +38,4 @@ public class PacketHandler {
         INSTANCE.sendToAllAround(packet, dim, x, y, z, distance);
     }
 
-    public static void sendRefreshPacket(TileHarvest tile) {
-        tile.hasChanged = true;
-        Packet<?> pkt = tile.getUpdatePacket();
-        if (pkt != null) {
-            WorldServer server = (WorldServer)tile.getWorld();
-            int chunkX = tile.getPos().getX() >> 4;
-            int chunkZ = tile.getPos().getZ() >> 4;
-            PlayerChunkMapEntry entry = server.getPlayerChunkMap().getEntry(chunkX, chunkZ);
-            if (entry != null) {
-                for (EntityPlayer player : tile.getWorld().playerEntities) {
-                    if (entry.containsPlayer((EntityPlayerMP) player)) {
-                        ((EntityPlayerMP) player).connection.sendPacket(pkt);
-                    }
-                }
-            }
-        }
-
-        tile.hasChanged = false;
-    }
 }