+++ src/main/java/joshie/harvest/npc/NPC.java
@@ -4,19 +4,15 @@ import joshie.harvest.api.buildings.Building;
 import joshie.harvest.api.buildings.BuildingLocation;
 import joshie.harvest.api.calendar.CalendarDate;
 import joshie.harvest.api.calendar.Season;
-import joshie.harvest.api.npc.IConditionalGreeting;
-import joshie.harvest.api.npc.INPC;
-import joshie.harvest.api.npc.INPCRegistry;
+import joshie.harvest.api.npc.*;
 import joshie.harvest.api.npc.INPCRegistry.Age;
 import joshie.harvest.api.npc.INPCRegistry.Gender;
-import joshie.harvest.api.npc.ISchedule;
 import joshie.harvest.api.npc.gift.IGiftHandler;
 import joshie.harvest.api.npc.gift.IGiftHandler.Quality;
 import joshie.harvest.api.shops.IShop;
 import joshie.harvest.cooking.HFCooking;
 import joshie.harvest.core.util.Text;
 import joshie.harvest.npc.entity.EntityNPC;
-import joshie.harvest.npc.greeting.GreetingMultiple;
 import joshie.harvest.npc.greeting.GreetingShop;
 import joshie.harvest.npc.item.ItemNPCTool.NPCTool;
 import joshie.harvest.shops.Shop;
@@ -36,7 +32,7 @@ import static joshie.harvest.core.lib.HFModInfo.*;
 public class NPC extends IForgeRegistryEntry.Impl<NPC> implements INPC {
     private final List<IConditionalGreeting> conditionals = new ArrayList<>(256);
     private final EnumMap<Location, BuildingLocation> locations;
-    private final GreetingMultiple npcGreetings;
+    private final String multipleLocalizationKey;
     private final String generalLocalizationKey;
     private final String localizationKey;
     private final ResourceLocation skin;
@@ -56,7 +52,7 @@ public class NPC extends IForgeRegistryEntry.Impl<NPC> implements INPC {
     private Shop shop;
     private boolean doesRespawn;
     private boolean alex;
-    private IConditionalGreeting infoGreeting;
+    private IGreeting infoGreeting;
     private ItemStack hasInfo;
 
     public NPC() {
@@ -77,7 +73,7 @@ public class NPC extends IForgeRegistryEntry.Impl<NPC> implements INPC {
         this.localizationKey = MODID + ".npc." + name + ".";
         this.generalLocalizationKey = MODID + ".npc.generic." + age.name().toLowerCase(Locale.ENGLISH) + ".";
         this.skin = new ResourceLocation(MODID, "textures/entity/" + name + ".png");
-        this.npcGreetings = new GreetingMultiple(name + ".greeting");
+        this.multipleLocalizationKey = MODID + ".npc." + name + ".greeting";
         this.locations = new EnumMap<>(Location.class);
         this.uuid = UUID.nameUUIDFromBytes(resource.toString().getBytes());
         this.setRegistryName(resource);
@@ -143,7 +139,7 @@ public class NPC extends IForgeRegistryEntry.Impl<NPC> implements INPC {
     }
 
     @Override
-    public INPC setHasInfo(ItemStack stack, IConditionalGreeting infoGreeting) {
+    public INPC setHasInfo(ItemStack stack, IGreeting infoGreeting) {
         this.hasInfo =stack;
         this.infoGreeting = infoGreeting;
         return this;
@@ -224,8 +220,8 @@ public class NPC extends IForgeRegistryEntry.Impl<NPC> implements INPC {
 
     @SuppressWarnings("unchecked")
     public String getInfoGreeting(EntityPlayer player, EntityNPC npc) {
-        if (infoGreeting == null || !infoGreeting.canDisplay(player, npc, this)) return null;
-        return infoGreeting.getLocalizedText(player, npc, npc.getNPC(), infoGreeting.getUnlocalizedText(player, npc, npc.getNPC()));
+        if (infoGreeting == null) return null;
+        return infoGreeting.getLocalizedText(player, npc, npc.getNPC());
     }
 
     public ResourceLocation getSkin() {
@@ -244,17 +240,22 @@ public class NPC extends IForgeRegistryEntry.Impl<NPC> implements INPC {
         return generalLocalizationKey;
     }
 
+    @Override
+    public ResourceLocation getResource() {
+        return getRegistryName();
+    }
+
     //Returns the script that this character should say at this point
+    @SuppressWarnings("unchecked")
     public String getGreeting(EntityPlayer player, EntityAgeable entity) {
         Collections.shuffle(conditionals);
         for (IConditionalGreeting greeting : conditionals) {
             if (greeting.canDisplay(player, entity, this) && player.worldObj.rand.nextDouble() * 100D < greeting.getDisplayChance()) {
-                if (greeting.getMaximumAlternatives() > 1) return Text.getRandomSpeech(player, entity, this, greeting, getRegistryName(), greeting.getUnlocalizedText(player, entity, this), greeting.getMaximumAlternatives());
-                else return greeting.getLocalizedText(player, entity, this, greeting.getUnlocalizedText(player, entity, this));
+                return greeting.getLocalizedText(player, entity, this);
             }
         }
 
-        return Text.getRandomSpeech(player, entity, this, npcGreetings, getRegistryName(), npcGreetings.getUnlocalizedText(player, entity, this), npcGreetings.getMaximumAlternatives());
+        return Text.getRandomSpeech(this, multipleLocalizationKey, 100);
     }
 
     @Override