+++ src/main/java/joshie/harvest/npc/greeting/GreetingShop.java
@@ -1,12 +1,13 @@
 package joshie.harvest.npc.greeting;
 
-import joshie.harvest.api.npc.IConditionalGreeting;
+import joshie.harvest.api.npc.IGreeting;
 import joshie.harvest.api.npc.INPC;
+import joshie.harvest.core.util.Text;
 import net.minecraft.entity.EntityAgeable;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.util.ResourceLocation;
 
-public class GreetingShop implements IConditionalGreeting {
+public class GreetingShop implements IGreeting {
     private final String text;
 
     public GreetingShop(ResourceLocation resourceLocation) {
@@ -14,7 +15,7 @@ public class GreetingShop implements IConditionalGreeting {
     }
 
     @Override
-    public String getUnlocalizedText(EntityPlayer player, EntityAgeable ageable, INPC npc) {
-        return text;
+    public String getLocalizedText(EntityPlayer player, EntityAgeable ageable, INPC npc) {
+        return Text.localize(text);
     }
 }