+++ src/main/java/joshie/harvest/plugins/immersiveengineering/StateHandlerHemp.java
@@ -7,12 +7,8 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IBlockAccess;
 
 public class StateHandlerHemp extends StateHandlerVanilla {
-    public StateHandlerHemp() {
-        super(null);
-    }
-
-    public void setBlock(Block block) {
-        this.block = block;
+    public StateHandlerHemp(Block block) {
+        super(block);
     }
 
     static int getMetaFromStage(int stage) {